requirements.txt
# requirements.txt

# Django
Django>=5.0

# Environment variables
django-environ>=0.8.1

# Document Processing (for AI analysis)
google-generativeai>=0.3.0
Pillow>=9.0.0
pdfminer.six>=20221105
python-docx>=0.8.11

# Forms
django-crispy-forms
crispy-bootstrap5

# HTTP Requests
requests

# Integrations
zoomus # For Zoom integration (keeping this)

# --- Google API for Gmail (Replacing Microsoft Graph) ---
google-api-python-client # Google API client library
google-auth # Google authentication library
google-auth-oauthlib # Google OAuth flow library

# Authentication
django-allauth # If you are using django-allauth

# File Storage (if using cloud storage like S3 or Google Cloud Storage)
django-storages # Example

# Add any other dependencies your project needs here
